moved validation rules to puzzle_validator.py. As a bonus I figured out why my Direction.__str__() wasn't working: I wasn't calling it...
