+++ srcs/main.py
@@ -3,13 +3,14 @@ import sys
 from srcs.npuzzle import Npuzzle
 from srcs.astar import Astar
 from srcs.heuristics import manhattan_distance
+from srcs.puzzle_validator import PuzzleValidator
 
 
 def main(npuzzle_file):
 	with open(npuzzle_file, 'r') as f:
 		puzzle = Npuzzle()
 		puzzle.parse_puzzle(f.read().splitlines())
-	if puzzle.is_solvable():
+	if PuzzleValidator.is_valid(puzzle) and PuzzleValidator.is_solvable(puzzle):
 		astar = Astar(puzzle, manhattan_distance)
 		astar.solve()
 	else:
