+++ srcs/npuzzle.py
@@ -3,7 +3,7 @@ import enum
 import numpy as np
 from typing import Tuple, List
 
-from srcs.parsing.parsing_file import parse_header, parserow, assert_validity, is_even, is_odd
+from srcs.parsing.parsing_file import parse_header, parserow
 
 
 class Direction(enum.IntEnum):
@@ -12,7 +12,7 @@ class Direction(enum.IntEnum):
 	DOWN = 2  # 0-pos changes with same pos in row below (zero-pos
 	LEFT = 3  # 0-pos changes with pos-1
 
-	def __str__(self):
+	def __str__(self) -> str:
 		return self.name
 
 
@@ -49,7 +49,6 @@ class Npuzzle:
 		self.rows = self.readrows(rows)
 		self.zero_pos = self.find_zero_pos()  # Tuple[xcoord, ycoord]
 		print(f'og is:\n{self.rows}\n\n')
-		assert_validity(self.size, self.rows)
 
 	def __lt__(self, other):
 		return self.moves < self.moves
@@ -116,7 +115,7 @@ class Npuzzle:
 
 	def extract_move_sequence(self) -> List[str]:
 		move_sequence = str(self.moves).replace('7', '')
-		return [Direction(int(move)) for move in move_sequence]
+		return [str(Direction(int(move))) for move in move_sequence]
 
 	def move_amount(self) -> int:
 		return len(str(self.moves)) - 1
@@ -130,18 +129,3 @@ class Npuzzle:
 		string += f'in {len(str(self.moves)) - 1} steps.\n'
 		string += f'Move sequence is {str(self.extract_move_sequence())}.\n'
 		return string
-
-	def is_solvable(self) -> bool:
-		flattened_puzzle = list(self.rows.flatten())
-		inversion_count = 0
-		print(flattened_puzzle)
-		for i in range(0, self.size ** 2 - 1):
-			for j in range(i + 1, self.size ** 2):
-				if flattened_puzzle[i] > flattened_puzzle[j] != 0 and flattened_puzzle[i] != 0:
-					inversion_count += 1
-		if is_odd(self.size) and is_even(inversion_count):
-			return True
-		elif is_even(self.size) and (is_even(self.size - self.zero_pos[0]) ^ is_odd(inversion_count)):
-			# the ^ operator is a XOR gate
-			return True
-		return False
